requirements.txt
asgiref==3.5.2
attrs==21.4.0
black==22.3.0
click==8.1.3
colorama==0.4.4
Django==4.0.5
django-debug-toolbar==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
inflection==0.5.1
jsonschema==4.6.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
